Moved usernames check to App model.
